@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.tunjid.fingergestures"
         minSdkVersion 26
-        targetSdkVersion 27
-        versionCode 38
-        versionName "6.0.2"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 40
+        versionName "7.0.0"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     signingConfigs {
         release {
@@ -57,36 +57,30 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.billingclient:billing:1.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.1'
-    implementation 'com.android.support:exifinterface:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:palette-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.palette:palette:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
     implementation 'com.google.android:flexbox:0.3.2'
+
+    implementation 'com.tunjid.android-bootstrap:core:3.0.1'
+    implementation 'com.tunjid.android-bootstrap:view:3.0.1'
 
     implementation "io.reactivex.rxjava2:rxjava:2.1.5"
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
 
     implementation 'com.squareup.picasso:picasso:2.5.2'
 
-    implementation('com.theartofdev.edmodo:android-image-cropper:2.6.0') {
+    implementation('com.theartofdev.edmodo:android-image-cropper:2.7.0') {
         exclude module: 'support-annotations'
         exclude module: 'recyclerview-v7'
         exclude module: 'exifinterface'
         exclude module: 'appcompat-v7'
         exclude module: 'support-v13'
         exclude module: 'support-v4'
-        exclude module: 'design'
-    }
-
-    implementation(group: "com.tunjid.android-bootstrap", name: "core", version: "2.0.0", changing: true) {
-        exclude module: 'appcompat-v7'
-        exclude module: 'support-annotations'
-        exclude module: 'support-v4'
-        exclude module: 'support-v13'
-        exclude module: 'recyclerview-v7'
         exclude module: 'design'
     }
 
@@ -100,6 +94,6 @@ dependencies {
     }
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.0-beta02'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-beta02'
 }
