@@ -1,6 +1,6 @@
 package com.tunjid.fingergestures;
 
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class WallpaperFileProvider extends FileProvider {}
