@@ -2,8 +2,8 @@ package com.tunjid.fingergestures.activities;
 
 import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.Window;
 
 import com.tunjid.fingergestures.BackgroundManager;
