@@ -2,8 +2,8 @@ package com.tunjid.fingergestures.gestureconsumers;
 
 import android.annotation.SuppressLint;
 import android.hardware.camera2.CameraManager;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.tunjid.fingergestures.App;
 
@@ -35,6 +35,7 @@ public class FlashlightGestureConsumer implements GestureConsumer {
         switch (gestureAction) {
             case TOGGLE_FLASHLIGHT:
                 App app = App.getInstance();
+                if (app == null) return;
                 if (!isCallbackRegistered) isCallbackRegistered = registerTorchCallback(app);
                 if (!isCallbackRegistered) return;
 
