@@ -1,16 +1,16 @@
 package com.tunjid.fingergestures.viewholders;
 
 import android.content.Context;
-import android.support.annotation.StringRes;
-import android.support.v7.app.AlertDialog;
+import androidx.annotation.StringRes;
+import androidx.appcompat.app.AlertDialog;
 import android.view.View;
 
-import com.tunjid.androidbootstrap.core.abstractclasses.BaseViewHolder;
+import com.tunjid.androidbootstrap.view.recyclerview.InteractiveViewHolder;
 import com.tunjid.fingergestures.R;
 import com.tunjid.fingergestures.adapters.AppAdapter;
 import com.tunjid.fingergestures.billing.PurchasesManager;
 
-public class AppViewHolder extends BaseViewHolder<AppAdapter.AppAdapterListener> {
+public class AppViewHolder extends InteractiveViewHolder<AppAdapter.AppAdapterListener> {
 
     public AppViewHolder(View itemView) {
         super(itemView);
