@@ -2,8 +2,8 @@ package com.tunjid.fingergestures.viewholders;
 
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.StringRes;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.StringRes;
 import android.view.View;
 import android.widget.TextView;
 
