@@ -2,17 +2,17 @@ package com.tunjid.fingergestures.viewholders;
 
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.tunjid.androidbootstrap.core.abstractclasses.BaseViewHolder;
+import com.tunjid.androidbootstrap.view.recyclerview.InteractiveViewHolder;
 import com.tunjid.fingergestures.R;
 import com.tunjid.fingergestures.adapters.PackageAdapter;
 
 
-public class PackageViewHolder extends BaseViewHolder<PackageAdapter.PackageClickListener> {
+public class PackageViewHolder extends InteractiveViewHolder<PackageAdapter.PackageClickListener> {
 
     private String packageName;
 
