@@ -1,27 +1,33 @@
 package com.tunjid.fingergestures.adapters;
 
-import android.support.v7.util.DiffUtil;
+import com.tunjid.androidbootstrap.view.recyclerview.InteractiveAdapter;
+import com.tunjid.androidbootstrap.view.recyclerview.InteractiveViewHolder;
 
-import com.tunjid.androidbootstrap.core.abstractclasses.BaseRecyclerViewAdapter;
-import com.tunjid.androidbootstrap.core.abstractclasses.BaseViewHolder;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.DiffUtil;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
+import androidx.recyclerview.widget.RecyclerView;
 import io.reactivex.Single;
+import io.reactivex.disposables.CompositeDisposable;
 
 
-public abstract class DiffAdapter<V extends BaseViewHolder<T>, T extends BaseRecyclerViewAdapter.AdapterListener> extends BaseRecyclerViewAdapter<V, T> {
+public abstract class DiffAdapter<V extends InteractiveViewHolder<T>, T extends InteractiveAdapter.AdapterListener> extends InteractiveAdapter<V, T> {
 
     final List<String> list;
     private final Supplier<List<String>> listSupplier;
+
+    private final CompositeDisposable disposables;
 
     DiffAdapter(Supplier<List<String>> listSupplier, T listener) {
         super(listener);
         setHasStableIds(true);
         this.list = new ArrayList<>();
         this.listSupplier = listSupplier;
+        this.disposables = new CompositeDisposable();
     }
 
     @Override
@@ -36,14 +42,19 @@ public abstract class DiffAdapter<V extends BaseViewHolder<T>, T extends BaseRec
 
     public void calculateDiff() {
         List<String> newPackageNames = listSupplier.get();
-        Single.fromCallable(() -> DiffUtil.calculateDiff(new DiffCallBack(list, newPackageNames)))
+        disposables.add(Single.fromCallable(() -> DiffUtil.calculateDiff(new DiffCallBack(list, newPackageNames)))
                 .subscribe(diffResult -> {
                     list.clear();
                     list.addAll(newPackageNames);
                     diffResult.dispatchUpdatesTo(this);
-                }, Throwable::printStackTrace);
+                }, Throwable::printStackTrace));
     }
 
+    @Override
+    public void onDetachedFromRecyclerView(@NonNull RecyclerView recyclerView) {
+        disposables.clear();
+        super.onDetachedFromRecyclerView(recyclerView);
+    }
 
     private static class DiffCallBack extends DiffUtil.Callback {
 
