@@ -3,12 +3,14 @@ package com.tunjid.fingergestures.adapters;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
 
-import com.tunjid.androidbootstrap.core.abstractclasses.BaseRecyclerViewAdapter;
+import com.tunjid.androidbootstrap.view.recyclerview.InteractiveAdapter;
 import com.tunjid.fingergestures.R;
 import com.tunjid.fingergestures.viewholders.PackageViewHolder;
 
 import java.util.List;
 import java.util.function.Supplier;
+
+import androidx.annotation.NonNull;
 
 
 public class PackageAdapter extends DiffAdapter<PackageViewHolder, PackageAdapter.PackageClickListener> {
@@ -21,18 +23,19 @@ public class PackageAdapter extends DiffAdapter<PackageViewHolder, PackageAdapte
         this.isHorizontal = isHorizontal;
     }
 
+    @NonNull
     @Override
-    public PackageViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public PackageViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         int layoutRes = isHorizontal ? R.layout.viewholder_package_horizontal : R.layout.viewholder_package_vertical;
         return new PackageViewHolder(LayoutInflater.from(parent.getContext()).inflate(layoutRes, parent, false), adapterListener);
     }
 
     @Override
-    public void onBindViewHolder(PackageViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull PackageViewHolder holder, int position) {
         holder.bind(list.get(position));
     }
 
-    public interface PackageClickListener extends BaseRecyclerViewAdapter.AdapterListener {
+    public interface PackageClickListener extends InteractiveAdapter.AdapterListener {
         void onPackageClicked(String packageName);
     }
 }
