@@ -3,17 +3,16 @@ package com.tunjid.fingergestures.gestureconsumers;
 import android.annotation.SuppressLint;
 import android.content.ContentResolver;
 import android.content.Context;
+import android.content.ContextWrapper;
 import android.content.Intent;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 import android.provider.Settings;
-import android.support.annotation.FloatRange;
-import android.support.annotation.IntRange;
-import android.support.v4.content.LocalBroadcastManager;
 
 import com.tunjid.fingergestures.App;
+import com.tunjid.fingergestures.BrightnessLookup;
 import com.tunjid.fingergestures.R;
 import com.tunjid.fingergestures.SetManager;
 import com.tunjid.fingergestures.activities.BrightnessActivity;
@@ -24,33 +23,39 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.function.Function;
 
+import androidx.annotation.FloatRange;
+import androidx.annotation.IntRange;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+
 import static android.provider.Settings.System.SCREEN_BRIGHTNESS;
 import static android.provider.Settings.System.SCREEN_BRIGHTNESS_MODE;
 import static android.provider.Settings.System.SCREEN_BRIGHTNESS_MODE_AUTOMATIC;
 import static android.provider.Settings.System.SCREEN_BRIGHTNESS_MODE_MANUAL;
-import static com.tunjid.fingergestures.App.requireApp;
-import static com.tunjid.fingergestures.gestureconsumers.GestureConsumer.normalizePercentageToByte;
+import static com.tunjid.fingergestures.App.transformApp;
+import static com.tunjid.fingergestures.App.withApp;
 import static com.tunjid.fingergestures.gestureconsumers.GestureConsumer.normalizePercentageToFraction;
 import static java.util.Comparator.naturalOrder;
 import static java.util.Comparator.reverseOrder;
 
 public class BrightnessGestureConsumer implements GestureConsumer {
 
-    static final float MAX_BRIGHTNESS = 255F;
+    private static final int FUZZ_THRESHOLD = 15;
     private static final float MIN_BRIGHTNESS = 0F;
     private static final float MIN_DIM_PERCENT = 0F;
-    private static final float MAX_DIM_PERCENT = 0.8F;
+    private static final float MAX_DIM_PERCENT = .8F;
+    private static final float MAX_BRIGHTNESS = 255F;
     private static final float DEF_DIM_PERCENT = MIN_DIM_PERCENT;
     private static final int MAX_ADAPTIVE_THRESHOLD = 1200;
     private static final int DEF_INCREMENT_VALUE = 20;
     private static final int DEF_POSITION_VALUE = 50;
     private static final int DEF_ADAPTIVE_BRIGHTNESS_THRESHOLD = 50;
 
-    public static final String BRIGHTNESS_FRACTION = "brightness value";
+    public static final String CURRENT_BRIGHTNESS_BYTE = "brightness value";
     public static final String ACTION_SCREEN_DIMMER_CHANGED = "show screen dimmer";
     private static final String INCREMENT_VALUE = "increment value";
     private static final String SLIDER_POSITION = "slider position";
     private static final String SLIDER_VISIBLE = "slider visible";
+    private static final String LOGARITHMIC_SCALE = "logarithmic scale";
     private static final String ADAPTIVE_BRIGHTNESS = "adaptive brightness";
     private static final String ADAPTIVE_BRIGHTNESS_THRESHOLD = "adaptive brightness threshold";
     private static final String SCREEN_DIMMER_ENABLED = "screen dimmer enabled";
@@ -59,7 +64,6 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     private static final String DISCRETE_BRIGHTNESS_SET = "discrete brightness values";
     private static final String EMPTY_STRING = "";
 
-    private final App app;
     private SetManager<Integer> discreteBrightnessManager;
 
     @SuppressLint("StaticFieldLeak")
@@ -71,8 +75,7 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     }
 
     private BrightnessGestureConsumer() {
-        app = App.getInstance();
-        discreteBrightnessManager = new SetManager<>(Integer::compareTo, s -> true, Integer::valueOf, String::valueOf);
+        discreteBrightnessManager = new SetManager<>(Integer::compareTo, ignored -> true, Integer::valueOf, String::valueOf);
     }
 
     @Override
@@ -80,7 +83,10 @@ public class BrightnessGestureConsumer implements GestureConsumer {
         int byteValue;
         int originalValue;
 
-        try {byteValue = Settings.System.getInt(app.getContentResolver(), SCREEN_BRIGHTNESS);}
+        App app = App.getInstance();
+        if (app == null) return;
+
+        try { byteValue = Settings.System.getInt(app.getContentResolver(), SCREEN_BRIGHTNESS); }
         catch (Exception e) {byteValue = (int) MAX_BRIGHTNESS;}
 
         originalValue = byteValue;
@@ -105,8 +111,7 @@ public class BrightnessGestureConsumer implements GestureConsumer {
 
         saveBrightness(byteValue);
 
-        float brightness = byteValue / MAX_BRIGHTNESS;
-        intent.putExtra(BRIGHTNESS_FRACTION, brightness);
+        intent.putExtra(CURRENT_BRIGHTNESS_BYTE, byteValue);
 
         if (shouldShowSlider()) app.startActivity(intent);
     }
@@ -128,7 +133,9 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     public void saveBrightness(int byteValue) {
         if (!App.canWriteToSettings()) return;
 
-        ContentResolver contentResolver = app.getContentResolver();
+        ContentResolver contentResolver = transformApp(ContextWrapper::getContentResolver);
+        if (contentResolver == null) return;
+
         Settings.System.putInt(contentResolver, SCREEN_BRIGHTNESS_MODE, SCREEN_BRIGHTNESS_MODE_MANUAL);
         Settings.System.putInt(contentResolver, SCREEN_BRIGHTNESS, byteValue);
     }
@@ -147,7 +154,7 @@ public class BrightnessGestureConsumer implements GestureConsumer {
             return;
         }
 
-        SensorManager sensorManager = (SensorManager) app.getSystemService(Context.SENSOR_SERVICE);
+        SensorManager sensorManager = transformApp(app -> (SensorManager) app.getSystemService(Context.SENSOR_SERVICE));
         if (sensorManager == null) return;
 
         final Sensor lightSensor = sensorManager.getDefaultSensor(Sensor.TYPE_LIGHT);
@@ -198,70 +205,87 @@ public class BrightnessGestureConsumer implements GestureConsumer {
 
     private int reduce(int byteValue) {
         return noDiscreteBrightness()
-                ? Math.max(byteValue - normalizePercentageToByte(getIncrementPercentage()), (int) MIN_BRIGHTNESS)
+                ? Math.max(adjustByte(byteValue, false), (int) MIN_BRIGHTNESS)
                 : findDiscreteBrightnessValue(byteValue, false);
     }
 
     private int increase(int byteValue) {
         return noDiscreteBrightness()
-                ? Math.min(byteValue + normalizePercentageToByte(getIncrementPercentage()), (int) MAX_BRIGHTNESS)
+                ? Math.min(adjustByte(byteValue, true), (int) MAX_BRIGHTNESS)
                 : findDiscreteBrightnessValue(byteValue, true);
     }
 
+    private int adjustByte(int byteValue, boolean increase) {
+        int delta = getIncrementPercentage();
+        int asPercentage = byteToPercentage(byteValue);
+        if (increase) asPercentage = Math.min(asPercentage + delta, 100);
+        else asPercentage = Math.max(asPercentage - delta, 0);
+
+        return percentToByte(asPercentage);
+    }
+
     public void setIncrementPercentage(@IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT) int incrementValue) {
-        app.getPreferences().edit().putInt(INCREMENT_VALUE, incrementValue).apply();
+        withApp(app -> app.getPreferences().edit().putInt(INCREMENT_VALUE, incrementValue).apply());
     }
 
     public void setPositionPercentage(@IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT) int positionPercentage) {
-        app.getPreferences().edit().putInt(SLIDER_POSITION, positionPercentage).apply();
+        withApp(app -> app.getPreferences().edit().putInt(SLIDER_POSITION, positionPercentage).apply());
     }
 
     private void setDimmerPercent(@FloatRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT) float percentage) {
-        app.getPreferences().edit().putFloat(SCREEN_DIMMER_DIM_PERCENT, percentage).apply();
+        withApp(app -> app.getPreferences().edit().putFloat(SCREEN_DIMMER_DIM_PERCENT, percentage).apply());
     }
 
     public void setAdaptiveBrightnessThreshold(@IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT) int threshold) {
-        app.getPreferences().edit().putInt(ADAPTIVE_BRIGHTNESS_THRESHOLD, threshold).apply();
+        withApp(app -> app.getPreferences().edit().putInt(ADAPTIVE_BRIGHTNESS_THRESHOLD, threshold).apply());
     }
 
     public void shouldRestoreAdaptiveBrightnessOnDisplaySleep(boolean restore) {
-        app.getPreferences().edit().putBoolean(ADAPTIVE_BRIGHTNESS, restore).apply();
+        withApp(app -> app.getPreferences().edit().putBoolean(ADAPTIVE_BRIGHTNESS, restore).apply());
+    }
+
+    public void shouldUseLogarithmicScale(boolean isLogarithmic) {
+        withApp(app -> app.getPreferences().edit().putBoolean(LOGARITHMIC_SCALE, isLogarithmic).apply());
     }
 
     public void setSliderVisible(boolean visible) {
-        app.getPreferences().edit().putBoolean(SLIDER_VISIBLE, visible).apply();
+        withApp(app -> app.getPreferences().edit().putBoolean(SLIDER_VISIBLE, visible).apply());
     }
 
     public void setAnimatesSlider(boolean visible) {
-        app.getPreferences().edit().putBoolean(ANIMATES_SLIDER, visible).apply();
+        withApp(app -> app.getPreferences().edit().putBoolean(ANIMATES_SLIDER, visible).apply());
     }
 
     public void setDimmerEnabled(boolean enabled) {
-        app.getPreferences().edit().putBoolean(SCREEN_DIMMER_ENABLED, enabled).apply();
+        withApp(app -> app.getPreferences().edit().putBoolean(SCREEN_DIMMER_ENABLED, enabled).apply());
         if (!enabled) removeDimmer();
     }
 
     @IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT)
     public int getIncrementPercentage() {
-        return app.getPreferences().getInt(INCREMENT_VALUE, DEF_INCREMENT_VALUE);
+        return transformApp(app -> app.getPreferences().getInt(INCREMENT_VALUE, DEF_INCREMENT_VALUE), DEF_INCREMENT_VALUE);
     }
 
     @IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT)
     public int getPositionPercentage() {
-        return app.getPreferences().getInt(SLIDER_POSITION, DEF_POSITION_VALUE);
+        return transformApp(app -> app.getPreferences().getInt(SLIDER_POSITION, DEF_POSITION_VALUE), DEF_POSITION_VALUE);
     }
 
     @IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT)
     public int getAdaptiveBrightnessThreshold() {
-        return app.getPreferences().getInt(ADAPTIVE_BRIGHTNESS_THRESHOLD, DEF_ADAPTIVE_BRIGHTNESS_THRESHOLD);
+        return transformApp(app -> app.getPreferences().getInt(ADAPTIVE_BRIGHTNESS_THRESHOLD, DEF_ADAPTIVE_BRIGHTNESS_THRESHOLD), DEF_ADAPTIVE_BRIGHTNESS_THRESHOLD);
     }
 
     public float getScreenDimmerDimPercent() {
-        return app.getPreferences().getFloat(SCREEN_DIMMER_DIM_PERCENT, DEF_DIM_PERCENT);
+        return transformApp(app -> app.getPreferences().getFloat(SCREEN_DIMMER_DIM_PERCENT, DEF_DIM_PERCENT), DEF_DIM_PERCENT);
     }
 
     public boolean restoresAdaptiveBrightnessOnDisplaySleep() {
-        return app.getPreferences().getBoolean(ADAPTIVE_BRIGHTNESS, false);
+        return transformApp(app -> app.getPreferences().getBoolean(ADAPTIVE_BRIGHTNESS, false), false);
+    }
+
+    public boolean usesLogarithmicScale() {
+        return transformApp(app -> app.getPreferences().getBoolean(LOGARITHMIC_SCALE, App.isPieOrHigher()), App.isPieOrHigher());
     }
 
     public boolean supportsAmbientThreshold() {
@@ -271,13 +295,13 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     }
 
     public boolean hasOverlayPermission() {
-        return Settings.canDrawOverlays(app);
+        return transformApp(Settings::canDrawOverlays, false);
     }
 
     public boolean isDimmerEnabled() {
         return hasOverlayPermission()
                 && PurchasesManager.getInstance().isPremium()
-                && app.getPreferences().getBoolean(SCREEN_DIMMER_ENABLED, false);
+                && transformApp(app -> app.getPreferences().getBoolean(SCREEN_DIMMER_ENABLED, false), false);
     }
 
     public boolean shouldShowDimmer() {
@@ -285,11 +309,11 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     }
 
     public boolean shouldShowSlider() {
-        return app.getPreferences().getBoolean(SLIDER_VISIBLE, true);
+        return transformApp(app -> app.getPreferences().getBoolean(SLIDER_VISIBLE, true), true);
     }
 
     public boolean shouldAnimateSlider() {
-        return app.getPreferences().getBoolean(ANIMATES_SLIDER, true);
+        return transformApp(app -> app.getPreferences().getBoolean(ANIMATES_SLIDER, true), true);
     }
 
     public boolean canAdjustDelta() {
@@ -305,7 +329,7 @@ public class BrightnessGestureConsumer implements GestureConsumer {
     }
 
     public String getAdjustDeltaText(int percentage) {
-        return requireApp(app -> PurchasesManager.getInstance().isPremium() && !noDiscreteBrightness()
+        return transformApp(app -> PurchasesManager.getInstance().isPremium() && !noDiscreteBrightness()
                 ? app.getString(R.string.delta_percent_premium, percentage)
                 : app.getString(R.string.delta_percent, percentage), EMPTY_STRING);
     }
@@ -316,52 +340,65 @@ public class BrightnessGestureConsumer implements GestureConsumer {
 
     public String getAdaptiveBrightnessThresholdText(@IntRange(from = GestureConsumer.ZERO_PERCENT, to = GestureConsumer.HUNDRED_PERCENT) int percent) {
         if (!hasBrightnessSensor())
-            return app.getString(R.string.unavailable_brightness_sensor);
+            return transformApp(app -> app.getString(R.string.unavailable_brightness_sensor), EMPTY_STRING);
 
         if (PurchasesManager.getInstance().isNotPremium())
-            return app.getString(R.string.go_premium_text);
+            return transformApp(app -> app.getString(R.string.go_premium_text), EMPTY_STRING);
 
         if (!restoresAdaptiveBrightnessOnDisplaySleep())
-            return (app.getString(R.string.adjust_adaptive_threshold_prompt));
+            return transformApp(app -> app.getString(R.string.adjust_adaptive_threshold_prompt), EMPTY_STRING);
 
         int lux = adaptiveThresholdToLux(percent);
-        String descriptor = app.getString(lux < 50
-                ? R.string.adaptive_threshold_low
-                : lux < 1000 ? R.string.adaptive_threshold_medium
-                : R.string.adaptive_threshold_high);
 
-        return app.getString(R.string.adaptive_threshold, lux, descriptor);
+        return transformApp(app -> {
+            String descriptor = app.getString(lux < 50
+                    ? R.string.adaptive_threshold_low
+                    : lux < 1000 ? R.string.adaptive_threshold_medium
+                    : R.string.adaptive_threshold_high);
+
+            return app.getString(R.string.adaptive_threshold, lux, descriptor);
+        }, EMPTY_STRING);
     }
 
     public void removeDimmer() {
         setDimmerPercent(MIN_DIM_PERCENT);
         Intent intent = new Intent(ACTION_SCREEN_DIMMER_CHANGED);
         intent.putExtra(SCREEN_DIMMER_DIM_PERCENT, getScreenDimmerDimPercent());
-        LocalBroadcastManager.getInstance(app).sendBroadcast(intent);
+        withApp(app -> LocalBroadcastManager.getInstance(app).sendBroadcast(intent));
+    }
+
+    public int percentToByte(int percentage) {
+        return usesLogarithmicScale()
+                ? BrightnessLookup.lookup(percentage, false)
+                : (int) (percentage * MAX_BRIGHTNESS / 100);
+    }
+
+    public int byteToPercentage(int byteValue) {
+        return usesLogarithmicScale()
+                ? BrightnessLookup.lookup(byteValue, true)
+                : (int) (byteValue * 100 / MAX_BRIGHTNESS);
     }
 
     private int adaptiveThresholdToLux(int percentage) {
         return (int) (percentage * MAX_ADAPTIVE_THRESHOLD / 100F);
     }
 
-    private int findDiscreteBrightnessValue(int current, boolean increasing) {
-        int evaluated = increasing ? current + 4 : current - 4;
+    private int findDiscreteBrightnessValue(int byteValue, boolean increasing) {
+        int evaluated = byteValue < FUZZ_THRESHOLD ? byteValue : increasing ? byteValue + 2 : byteValue - 2;
         int alternative = (int) (increasing ? MAX_BRIGHTNESS : MIN_BRIGHTNESS);
         Comparator<Integer> comparator = increasing ? naturalOrder() : reverseOrder();
         Function<Integer, Boolean> filter = integer -> increasing ? integer > evaluated : integer < evaluated;
 
         return discreteBrightnessManager.getSet(DISCRETE_BRIGHTNESS_SET).stream()
-                .map(BrightnessGestureConsumer::stringPercentageToBrightnessInt)
-                .filter(filter::apply)
-                .sorted(comparator)
-                .findFirst().orElse(alternative);
+                .map(this::stringPercentageToBrightnessInt)
+                .filter(filter::apply).min(comparator).orElse(alternative);
     }
 
     private void toggleAdaptiveBrightness(boolean on) {
         int brightnessMode = on
                 ? SCREEN_BRIGHTNESS_MODE_AUTOMATIC
                 : SCREEN_BRIGHTNESS_MODE_MANUAL;
-        Settings.System.putInt(app.getContentResolver(), SCREEN_BRIGHTNESS_MODE, brightnessMode);
+        withApp(app -> Settings.System.putInt(app.getContentResolver(), SCREEN_BRIGHTNESS_MODE, brightnessMode));
     }
 
     private static float roundDown(float d) {
@@ -370,14 +407,13 @@ public class BrightnessGestureConsumer implements GestureConsumer {
         return bd.floatValue();
     }
 
-    private static int stringPercentageToBrightnessInt(String stringPercent) {
+    private int stringPercentageToBrightnessInt(String stringPercent) {
         Float percentage = Float.valueOf(stringPercent);
-        Float byteFloat = percentage * MAX_BRIGHTNESS / 100;
-        return byteFloat.intValue();
+        return percentToByte(percentage.intValue());
     }
 
     private boolean hasBrightnessSensor() {
-        SensorManager sensorManager = (SensorManager) app.getSystemService(Context.SENSOR_SERVICE);
+        SensorManager sensorManager = transformApp(app -> (SensorManager) app.getSystemService(Context.SENSOR_SERVICE));
         if (sensorManager == null) return false;
 
         final Sensor lightSensor = sensorManager.getDefaultSensor(Sensor.TYPE_LIGHT);
